requirements.txt
Flask==1.1.1
BeautifulSoup4==4.8.0
requests==2.22.0
Pillow==6.2.1
numpy==1.17.4
tensorflow==2.0.0